pyproject.toml
[project]
name = "transcript-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
]
